Adapted SDReconfigurator.
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Incubation/SimuLizar/branches/Elasticity%20and%20RR%20model%20merge@31612 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ org.palladiosimulator.simulizar/src/org/palladiosimulator/simulizar/runtimestate/SimulatedBasicComponentInstance.java
@@ -27,7 +27,7 @@ public class SimulatedBasicComponentInstance extends SimulatedComponentInstance 
         this.passiveResourcesMap = new HashMap<String, IPassiveResource>();
         final AssemblyContext myAssCtx = fqID.getAssembyContextPath().get(fqID.getAssembyContextPath().size() - 1);
         for (final PassiveResource passiveResource : passiveResources) {
-            final long initialCount = StackContext.evaluateStatic(
+            final long initialCount = (long) StackContext.evaluateStatic(
                     passiveResource.getCapacity_PassiveResource().getSpecification(), Integer.class,
                     context.getStack().currentStackFrame());
             final IPassiveResource simulatedResource = new SimSimpleFairPassiveResource(passiveResource, myAssCtx,
